refactor(lib, tool): Move type checks to library
Move the custom type checks we used in the tool into the library and add
unit tests to everything.

One interesting thing about this is that the tests need access to
ASP.NET Core in order to have the correct types available. This means
that the test project includes a framework reference to
Microsoft.AspNetCore.App and can technically be started as a web
project. This will however fail miserably.